+++ libseccomp/src/api.rs
@@ -3,7 +3,7 @@
 // Copyright 2021 Sony Group Corporation
 //
 
-use crate::error::ErrorKind::*;
+use super::cvt;
 use crate::error::{Result, SeccompError};
 use crate::version::ensure_supported_version;
 use crate::{check_version, ScmpVersion};
@@ -28,39 +28,21 @@ use libseccomp_sys::*;
 /// If the API level can not be detected due to the library being older than v2.4.0,
 /// an error will be returned.
 pub fn set_api(level: u32) -> Result<()> {
-    let ret = unsafe { seccomp_api_set(level) };
-    if ret != 0 {
-        return Err(SeccompError::new(Common(
-            "API level operations are not supported".to_string(),
-        )));
-    }
+    cvt(unsafe { seccomp_api_set(level) })?;
 
     Ok(())
 }
 
 /// Gets the API level supported by the system.
 ///
-/// This function returns a positive int containing the API level.
 /// See the [`seccomp_api_get(3)`] man page for details on available API levels.
 ///
 /// [`seccomp_api_get(3)`]: https://www.man7.org/linux/man-pages/man3/seccomp_api_get.3.html
 ///
 /// This function corresponds to
 /// [`seccomp_api_get`](https://www.man7.org/linux/man-pages/man3/seccomp_api_get.3.html).
-///
-/// # Errors
-///
-/// If the API level can not be detected due to the library being older than v2.4.0,
-/// an error will be returned.
-pub fn get_api() -> Result<u32> {
-    let ret = unsafe { seccomp_api_get() };
-    if ret == 0 {
-        return Err(SeccompError::new(Common(
-            "API level operations are not supported".to_string(),
-        )));
-    }
-
-    Ok(ret)
+pub fn get_api() -> u32 {
+    unsafe { seccomp_api_get() }
 }
 
 /// Checks that both the libseccomp API level and the libseccomp version being
@@ -80,7 +62,7 @@ pub fn get_api() -> Result<u32> {
 /// If an issue is encountered getting the current API level or version,
 /// an error will be returned.
 pub fn check_api(min_level: u32, expected: ScmpVersion) -> Result<bool> {
-    let level = get_api()?;
+    let level = get_api();
 
     if level >= min_level && check_version(expected)? {
         Ok(true)
@@ -104,16 +86,16 @@ pub fn check_api(min_level: u32, expected: ScmpVersion) -> Result<bool> {
 /// If the libseccomp API level and the libseccomp version being used are less than
 /// the specified version, an error will be returned.
 pub(crate) fn ensure_supported_api(msg: &str, min_level: u32, expected: ScmpVersion) -> Result<()> {
-    let level = get_api()?;
+    let level = get_api();
 
     if level >= min_level {
         ensure_supported_version(msg, expected)
     } else {
         let current = ScmpVersion::current()?;
-        Err(SeccompError::new(Common(format!(
+        Err(SeccompError::with_msg(format!(
             "{} requires libseccomp >= {} and API level >= {} (current version: {}, API level: {})",
             msg, expected, min_level, current, level
-        ))))
+        )))
     }
 }